+++ lib/modules/core/core_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: core
-// Updated at: 2026-04-04 18:27:25
-// Commit: fix(core): fix memory leak by cancelling core subscription
+// Updated at: 2026-04-04 21:03:18
+// Commit: perf(core): apply RepaintBoundary on animated core widgets
 
 class coreComponentState {
-  final String commitRef = '122';
-  final String lastMessage = 'fix(core): fix memory leak by cancelling core subscription';
-  final DateTime timestamp = DateTime.parse('2026-04-04T18:27:25+05:00');
+  final String commitRef = '131';
+  final String lastMessage = 'perf(core): apply RepaintBoundary on animated core widgets';
+  final DateTime timestamp = DateTime.parse('2026-04-04T21:03:18+05:00');
 }