+++ CHANGELOG.md
@@ -132,3 +132,4 @@
 - [2026-04-04 20:30] fix(state): resolve race condition during state initialization
 - [2026-04-04 20:37] test(network): add unit tests for network business logic
 - [2026-04-04 20:59] fix(state): fix memory leak by cancelling state subscription
+- [2026-04-04 21:03] perf(core): apply RepaintBoundary on animated core widgets